requirements.txt
aiohttp
aiofiles
pydantic
pydantic-settings
python-dotenv